+++ src/components/LoadingScreen/index.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from 'react'
 
 import { useThreeContext } from '../../contexts/ThreeJSContext'
+import { useUIContext } from '../../contexts/UIContext'
 import Loader from '../../three/experience/Loader'
 import { Container } from './styles'
 
 const LoadingScreen: React.FC = () => {
   const { threeExperience } = useThreeContext()
-
-  const [show, setShow] = useState(true)
+  const { setLoading } = useUIContext()
 
   const [loader, setLoader] = useState<Loader>()
   const [progress, setProgress] = useState(0)
@@ -17,16 +17,16 @@ const LoadingScreen: React.FC = () => {
     const loaderInstance = new Loader(
       threeExperience,
       (p) => setProgress(p),
-      () => setShow(false)
+      () => setLoading(false)
     )
     setLoader(loaderInstance)
   }, [threeExperience, loader])
 
-  return show ? (
+  return (
     <Container className={progress === 1 ? 'loaded' : ''}>
       <h2>{Math.round(progress * 10000) / 100}%</h2>
     </Container>
-  ) : null
+  )
 }
 
 export default LoadingScreen
