+++ src/components/Overlay/styles.ts
@@ -35,67 +35,6 @@ export const Container = styled.div`
     display: none;
   }
 
-  .blocker,
-  .portrait {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    background-color: rgba(0, 0, 0, 0.8);
-    color: white;
-
-    h2 {
-      font-size: 2rem;
-    }
-    p,
-    span {
-      font-size: 1.5rem;
-    }
-
-    .instructions {
-      width: 100%;
-      height: 100%;
-
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-
-      text-align: center;
-      font-size: 14px;
-
-      pointer-events: all;
-      cursor: pointer;
-    }
-  }
-
-  .portrait {
-    display: none;
-  }
-
-  @media (orientation: portrait) {
-    .portrait {
-      display: flex;
-      pointer-events: all;
-      flex-direction: column;
-      justify-content: space-evenly;
-      align-items: center;
-      text-align: center;
-      padding: 20px;
-
-      button {
-        outline: none;
-        border: none;
-        background-color: white;
-        border-radius: 10px;
-        width: 50%;
-        height: 50px;
-        font-size: 1.2rem;
-      }
-    }
-  }
-
   .background {
     position: absolute;
     width: 100%;
